Crate replicapool1_beta2

Features

Handle the following Resources with ease from the central hub ...

instance group managers (abandoninstances, delete, deleteinstances, get, insert, list, recreateinstances, resize, setinstancetemplate and settargetpools)

zone operations (get and list)

Everything else about the replicapool API can be found at the official documentation site.

Not what you are looking for ? Find all other google APIs in their Rust documentation index.

Structure of this Library

The API is structured into the following primary items:

Hub

a central object to maintain state and allow accessing all Activities

Resources

primary types that you can apply Activities to

a collection of properties and Parts

Parts

a collection of properties

never directly used in Activities

Activities

operations to apply to Resources

Generally speaking, you can invoke Activities like this:

let r = hub.resource().activity(...).doit()

Or specifically ...

let r = hub.instance_group_managers().set_target_pools(...).doit()
let r = hub.instance_group_managers().list(...).doit()
let r = hub.instance_group_managers().insert(...).doit()
let r = hub.instance_group_managers().get(...).doit()
let r = hub.instance_group_managers().abandon_instances(...).doit()
let r = hub.instance_group_managers().recreate_instances(...).doit()
let r = hub.instance_group_managers().delete(...).doit()
let r = hub.instance_group_managers().set_instance_template(...).doit()
let r = hub.instance_group_managers().resize(...).doit()
let r = hub.instance_group_managers().delete_instances(...).doit()

The resource() and activity(...) calls create builders. The second one dealing with Activities supports various methods to configure the impending operation (not shown here). It is made such that all required arguments have to be specified right away (i.e. (...)), whereas all optional ones can be build up as desired. The doit() method performs the actual communication with the server and returns the respective result.

Usage

A complete example

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate replicapool1_beta2;
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
 
// Get an ApplicationSecret instance by some means. It contains the `client_id` and `client_secret`, 
// among other things.
let secret: ApplicationSecret = Default::default();
// Instantiate the authenticator. It will choose a suitable authentication flow for you, 
// unless you replace  `None` with the desired Flow
// Provide your own `AuthenticatorDelegate` to adjust the way it operates and get feedback about what's going on
// You probably want to bring in your own `TokenStorage` to persist tokens and retrieve them from storage.
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Replicapool::new(hyper::Client::new(), auth);
// You can configure optional parameters by calling the respective setters at will, and
// execute the final call using `doit()`.
// Values shown here are possibly random and not representative !
let result = hub.instance_group_managers().list("project", "zone")
             .page_token("accusam")
             .max_results(93)
             .filter("justo")
             .doit();
 
match result {
    Result::HttpError(err) => println!("HTTPERROR: {:?}", err),
    Result::FieldClash(clashed_field) => println!("FIELD CLASH: {:?}", clashed_field),
    Result::Success(value) => println!("Result Value: {:?}", value),
}

Handling Errors

All errors produced by the system are provided either as Result enumeration as return value of the doit() methods, or handed as possibly intermediate results to either the Hub Delegate, or the Authenticator Delegate.

When delegates handle errors or intermediate values, they may have a chance to instruct the system to retry. This makes the system potentially resilient to all kinds of errors.

About Customization/Callbacks

You may alter the way an doit() method is called by providing a delegate to the Method Builder before making the final doit() call. Respective methods will be called to provide progress information, as well as determine whether the system should retry on failure.

The delegate trait is default-implemented, allowing you to customize it with minimal effort.

About Parts

All structures provided by this library are made to be enocodable and decodable via json. Optionals are used to indicate that partial requests are responses are valid. Most optionals are are considered Parts which are identifyable by name, which will be sent to the server to indicate either the set parts of the request or the desired parts in the response.

About Builder Arguments

Using method builders, you are able to prepare an action call by repeatedly calling it's methods. These will always take a single argument, for which the following statements are true.

PODs are handed by copy

strings are passed as &str

request values are borrowed

Arguments will always be copied or cloned into the builder, to make them independent of their original life times.

Modules

cmn

Macros

map!
This macro is advertised in the documentation, which is why we deliver it as well

Structs

InstanceGroupManager
An Instance Group Manager resource.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

setTargetPools (none)

list (none)

insert (request)

get (response)

abandonInstances (none)

recreateInstances (none)

delete (none)

setInstanceTemplate (none)

resize (none)

deleteInstances (none)

InstanceGroupManagerAbandonInstanceMethodBuilder
Removes the specified instances from the managed instance group, and from any target pools of which they were members, without deleting the instances.

InstanceGroupManagerDeleteInstanceMethodBuilder
Deletes the specified instances. The instances are removed from the instance group and any target pools of which they are a member, then deleted. The targetSize of the instance group manager is reduced by the number of instances deleted.

InstanceGroupManagerDeleteMethodBuilder
Deletes the instance group manager and all instances contained within. If you'd like to delete the manager without deleting the instances, you must first abandon the instances to remove them from the group.

InstanceGroupManagerGetMethodBuilder
Returns the specified Instance Group Manager resource.

InstanceGroupManagerInsertMethodBuilder
Creates an instance group manager, as well as the instance group and the specified number of instances.

InstanceGroupManagerList
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list (response)

InstanceGroupManagerListMethodBuilder
Retrieves the list of Instance Group Manager resources contained within the specified zone.

InstanceGroupManagerMethodsBuilder
A builder providing access to all methods supported on instanceGroupManager resources. It is not used directly, but through the Replicapool hub.

InstanceGroupManagerRecreateInstanceMethodBuilder
Recreates the specified instances. The instances are deleted, then recreated using the instance group manager's current instance template.

InstanceGroupManagerResizeMethodBuilder
Resizes the managed instance group up or down. If resized up, new instances are created using the current instance template. If resized down, instances are removed in the order outlined in Resizing a managed instance group.

InstanceGroupManagerSetInstanceTemplateMethodBuilder
Sets the instance template to use when creating new instances in this group. Existing instances are not affected.

InstanceGroupManagerSetTargetPoolMethodBuilder
Modifies the target pools to which all new instances in this group are assigned. Existing instances in the group are not affected.

InstanceGroupManagersAbandonInstancesRequest
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

abandonInstances (request)

InstanceGroupManagersDeleteInstancesRequest
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

deleteInstances (request)

InstanceGroupManagersRecreateInstancesRequest
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

recreateInstances (request)

InstanceGroupManagersSetInstanceTemplateRequest
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

setInstanceTemplate (request)

InstanceGroupManagersSetTargetPoolsRequest
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

setTargetPools (request)

Operation
An operation resource, used to manage asynchronous API requests.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

setTargetPools (response)

insert (response)

abandonInstances (response)

recreateInstances (response)

delete (response)

setInstanceTemplate (response)

get (response)

resize (response)

deleteInstances (response)

OperationError
[Output Only] If errors occurred during processing of this operation, this field will be populated.

This type is not used in any activity, and only used as part of another schema.

OperationErrorErrors
[Output Only] The array of errors encountered while processing this operation.

This type is not used in any activity, and only used as part of another schema.

OperationList
There is no detailed description.

Activities

This type is used in activities, which are methods you may call on this type or where this type is involved in. The list links the activity name, along with information about where it is used (one of request and response).

list (response)

OperationWarnings
[Output Only] If there are issues with this operation, a warning is returned.

This type is not used in any activity, and only used as part of another schema.

OperationWarningsData
[Output only] Metadata for this warning in key:value format.

This type is not used in any activity, and only used as part of another schema.

Replicapool
Central instance to access all Replicapool related resource activities

ZoneOperationGetMethodBuilder
Retrieves the specified zone-specific operation resource.

ZoneOperationListMethodBuilder
Retrieves the list of operation resources contained within the specified zone.

ZoneOperationMethodsBuilder
A builder providing access to all methods supported on zoneOperation resources. It is not used directly, but through the Replicapool hub.

Enums

Scope
Identifies the an OAuth2 authorization scope. A scope is needed when requesting an authorization token.