Struct replicapool1_beta2::InstanceGroupManagerMethodsBuilder

pub struct InstanceGroupManagerMethodsBuilder<'a, C, NC, A> where NC: 'a, C: 'a, A: 'a {
    // some fields omitted
}

A builder providing access to all methods supported on instanceGroupManager resources. It is not used directly, but through the Replicapool hub.

Example

Instantiate a resource builder

extern crate hyper;
extern crate "yup-oauth2" as oauth2;
extern crate "rustc-serialize" as rustc_serialize;
extern crate replicapool1_beta2;
 
use std::default::Default;
use oauth2::{Authenticator, DefaultAuthenticatorDelegate, ApplicationSecret, MemoryStorage};
use replicapool1_beta2::Replicapool;
 
let secret: ApplicationSecret = Default::default();
let auth = Authenticator::new(&secret, DefaultAuthenticatorDelegate,
                              hyper::Client::new(),
                              <MemoryStorage as Default>::default(), None);
let mut hub = Replicapool::new(hyper::Client::new(), auth);
// Usually you wouldn't bind this to a variable, but keep calling *MethodBuilders*
// like `abandon_instances(...)`, `delete(...)`, `delete_instances(...)`, `get(...)`, `insert(...)`, `list(...)`, `recreate_instances(...)`, `resize(...)`, `set_instance_template(...)` and `set_target_pools(...)`
// to build up your call.
let rb = hub.instance_group_managers();

Methods

impl<'a, C, NC, A> InstanceGroupManagerMethodsBuilder<'a, C, NC, A>

fn set_target_pools(&self, request: &InstanceGroupManagersSetTargetPoolsRequest, project: &str, zone: &str, instance_group_manager: &str) -> InstanceGroupManagerSetTargetPoolMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Modifies the target pools to which all new instances in this group are assigned. Existing instances in the group are not affected.

fn list(&self, project: &str, zone: &str) -> InstanceGroupManagerListMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Retrieves the list of Instance Group Manager resources contained within the specified zone.

fn insert(&self, request: &InstanceGroupManager, project: &str, zone: &str, size: i32) -> InstanceGroupManagerInsertMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Creates an instance group manager, as well as the instance group and the specified number of instances.

fn get(&self, project: &str, zone: &str, instance_group_manager: &str) -> InstanceGroupManagerGetMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Returns the specified Instance Group Manager resource.

fn abandon_instances(&self, request: &InstanceGroupManagersAbandonInstancesRequest, project: &str, zone: &str, instance_group_manager: &str) -> InstanceGroupManagerAbandonInstanceMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Removes the specified instances from the managed instance group, and from any target pools of which they were members, without deleting the instances.

fn recreate_instances(&self, request: &InstanceGroupManagersRecreateInstancesRequest, project: &str, zone: &str, instance_group_manager: &str) -> InstanceGroupManagerRecreateInstanceMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Recreates the specified instances. The instances are deleted, then recreated using the instance group manager's current instance template.

fn delete(&self, project: &str, zone: &str, instance_group_manager: &str) -> InstanceGroupManagerDeleteMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Deletes the instance group manager and all instances contained within. If you'd like to delete the manager without deleting the instances, you must first abandon the instances to remove them from the group.

fn set_instance_template(&self, request: &InstanceGroupManagersSetInstanceTemplateRequest, project: &str, zone: &str, instance_group_manager: &str) -> InstanceGroupManagerSetInstanceTemplateMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Sets the instance template to use when creating new instances in this group. Existing instances are not affected.

fn resize(&self, project: &str, zone: &str, instance_group_manager: &str, size: i32) -> InstanceGroupManagerResizeMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Resizes the managed instance group up or down. If resized up, new instances are created using the current instance template. If resized down, instances are removed in the order outlined in Resizing a managed instance group.

fn delete_instances(&self, request: &InstanceGroupManagersDeleteInstancesRequest, project: &str, zone: &str, instance_group_manager: &str) -> InstanceGroupManagerDeleteInstanceMethodBuilder<'a, C, NC, A>

Create a builder to help you perform the following task:

Deletes the specified instances. The instances are removed from the instance group and any target pools of which they are a member, then deleted. The targetSize of the instance group manager is reduced by the number of instances deleted.